feat(trusty-review): forced structured output (Bedrock tool-use + OpenRouter json_schema) to eliminate fail-safe verdicts
- Add `ResponseSchema { name, schema }` to `LlmRequest` (optional); when set,
  providers MUST force the model to emit JSON conforming to the schema and
  return only the clean JSON object in `LlmResponse.text`.
- Bedrock: build a `ToolConfiguration` with a single tool whose `inputSchema`
  is the caller's JSON Schema and `toolChoice = TOOL(name)` (forced); extract
  `toolUse.input` as the response text.  New `bedrock/tool_use.rs` module.
- OpenRouter: send `response_format: { type: "json_schema", json_schema: { name,
  strict: true, schema } }` when schema is present (OpenAI-compatible structured
  outputs, passed through by OpenRouter).
- Wire all three callers to include their schema:
  - `pipeline/prompt.rs`: `review_output` schema (verdict/summary/findings).
  - `profile/batch_reviewer.rs`: `period_findings` schema.
  - `profile/synthesizer_prompt.rs`: `synthesis_output` schema.
- `parse_review_response` gains Strategy 1: direct JSON parse (succeeds for
  structured output); fence-based and keyword-scan remain as fallbacks.
  Fail-safe APPROVE is now error-only, not a parse failure.
- Swap default compare-set: drop Opus (not access-granted in target account);
  add confirmed-available Sonnet 4.5 (`us.anthropic.claude-sonnet-4-5-20250929-v1:0`).
  Add Sonnet 4.5 pricing ($3/M in, $15/M out) to Bedrock pricing table.
- Split old 784-line `bedrock.rs` into `bedrock/` submodule:
  `mod.rs` (443L), `pricing.rs` (122L), `tool_use.rs` (316L), `tests.rs` (391L).
- Extract test files for openrouter, parser, prompt to stay under 500-line cap.
