+++ crates/trusty-review/src/pipeline/prompt.rs
@@ -26,7 +26,7 @@ use crate::{
         analyze_client::{ComplexityHotspot, Smell},
         search_client::SearchResult,
     },
-    llm::{ChatMessage, LlmRequest, strip_provider_prefix},
+    llm::{ChatMessage, LlmRequest, ResponseSchema, strip_provider_prefix},
     models::ReviewResult,
 };
 
@@ -37,6 +37,61 @@ const REVIEWER_TEMPERATURE: f32 = 0.3;
 
 /// Maximum tokens for the review response.
 const REVIEWER_MAX_TOKENS: u32 = 4096;
+
+// ─── Review output schema ─────────────────────────────────────────────────────
+
+/// The name used for the structured-output tool/schema.
+const REVIEW_SCHEMA_NAME: &str = "review_output";
+
+/// Build the JSON Schema for the review output structure.
+///
+/// Why: the provider uses this schema to force the model to emit a clean JSON
+/// object rather than free text with a JSON block embedded in it.  This
+/// eliminates the fail-safe APPROVE problem (Haiku always fail-safes; Sonnet
+/// sometimes does) that occurs when the model ignores the output format
+/// instruction in the system prompt.
+/// What: returns a `ResponseSchema` whose `schema` field is a JSON Schema
+/// object describing the `review_output` shape expected by `parse_review_response`.
+/// The schema matches the fields that `LlmOutputBlock` deserializes.
+/// Test: `build_review_prompt_includes_response_schema` in this module.
+pub fn review_response_schema() -> ResponseSchema {
+    ResponseSchema {
+        name: REVIEW_SCHEMA_NAME.to_string(),
+        schema: serde_json::json!({
+            "type": "object",
+            "properties": {
+                "verdict": {
+                    "type": "string",
+                    "enum": ["APPROVE", "APPROVE*", "REQUEST_CHANGES", "BLOCK", "N/A"],
+                    "description": "Review verdict"
+                },
+                "summary": {
+                    "type": "string",
+                    "description": "One-line summary of the review"
+                },
+                "findings": {
+                    "type": "array",
+                    "items": {
+                        "type": "object",
+                        "properties": {
+                            "title": {"type": "string"},
+                            "body": {"type": "string"},
+                            "severity": {
+                                "type": "string",
+                                "enum": ["low", "medium", "high", "critical"]
+                            },
+                            "confidence": {"type": "number", "minimum": 0.0, "maximum": 1.0},
+                            "file": {"type": "string"},
+                            "line": {"type": ["integer", "null"]}
+                        },
+                        "required": ["title", "body"]
+                    }
+                }
+            },
+            "required": ["verdict", "summary", "findings"]
+        }),
+    }
+}
 
 // ─── Context inputs ───────────────────────────────────────────────────────────
 
@@ -66,6 +121,11 @@ pub struct ReviewContext {
 /// Keeping it as a function (not a constant) allows conditional sections in the
 /// future (e.g. copilot-mode conditioning from spec REV-104).
 /// What: returns a static string; future versions may accept a context param.
+/// The output-format section no longer instructs the model to emit a fenced
+/// JSON block — with forced structured output enabled (via `response_schema`),
+/// the provider guarantees the response IS the JSON object, so fence-based
+/// instructions are unnecessary and would confuse models that try to
+/// literally wrap output in backticks.
 /// Test: `system_prompt_contains_policy`.
 pub fn reviewer_system_prompt() -> &'static str {
     r#"You are a senior software engineer performing a pull-request code review.
@@ -84,32 +144,17 @@ pub fn reviewer_system_prompt() -> &'static str {
 Focus on: correctness bugs, security issues, data-loss risks, logic errors.
 Note but do not block on: style, minor naming, documentation gaps, test coverage.
 
-## Output format (REQUIRED)
-After your review narrative, output EXACTLY ONE JSON block as your final content,
-using the following schema. Do not include any text after the JSON block.
+## Output (REQUIRED)
+Populate the structured response with:
+- `verdict`: one of APPROVE, APPROVE*, REQUEST_CHANGES, BLOCK, N/A.
+- `summary`: one sentence summary of the review.
+- `findings`: array of issues found (empty array if none).
+  Each finding has: title, body (detailed description), severity (low/medium/high/critical),
+  confidence (0.0–1.0), file (source file path), line (null if not applicable).
 
-```json
-{
-  "verdict": "APPROVE",
-  "summary": "One sentence summary of the review.",
-  "findings": [
-    {
-      "title": "Short finding title",
-      "body": "Detailed description of the issue.",
-      "severity": "low|medium|high|critical",
-      "confidence": 0.85,
-      "file": "src/path/to/file.rs",
-      "line": 42
-    }
-  ]
-}
-```
-
-`verdict` must be one of: APPROVE, APPROVE*, REQUEST_CHANGES, BLOCK, N/A.
 `confidence` is a float in [0.0, 1.0].
 `line` may be null if no specific line is applicable.
-`findings` may be an empty array if there are no issues.
-Emit the raw JSON block with no additional prose after it."#
+`findings` may be an empty array if there are no issues."#
 }
 
 // ─── Prompt builder ───────────────────────────────────────────────────────────
@@ -120,12 +165,15 @@ Emit the raw JSON block with no additional prose after it."#
 /// prompt iteration doesn't require touching pipeline logic.
 /// What: assembles a system prompt + user message containing the PR metadata,
 /// truncated diff, code search context (if any), and static-analysis annotations
-/// (if any).  The user message ends with the structured-output reminder.
+/// (if any).  Includes `response_schema` so the provider forces structured
+/// output via Bedrock tool-use or OpenRouter json_schema — eliminating the
+/// silent fail-safe APPROVE problem.
 /// `reviewer_model` may carry a `bedrock/` or `openrouter/` routing prefix;
 /// this function strips it before setting `LlmRequest.model` so the bare id is
 /// what reaches the provider's API (the prefix is used only for routing).
 /// Test: `build_review_prompt_includes_diff`, `prompt_includes_context_blocks`,
-/// `build_review_prompt_strips_bedrock_prefix`.
+/// `build_review_prompt_strips_bedrock_prefix`,
+/// `build_review_prompt_includes_response_schema`.
 pub fn build_review_prompt(
     owner: &str,
     repo: &str,
@@ -144,6 +192,7 @@ pub fn build_review_prompt(
         }],
         temperature: REVIEWER_TEMPERATURE,
         max_tokens: REVIEWER_MAX_TOKENS,
+        response_schema: Some(review_response_schema()),
     }
 }
 
@@ -266,10 +315,10 @@ fn build_user_message(
         msg.push('\n');
     }
 
-    // Structured-output reminder.
+    // Structured-output instruction (schema-enforced; no need to emit a fence).
     msg.push_str(
-        "Please review the diff above and end your response with the structured \
-         JSON verdict block exactly as specified in the system prompt.\n",
+        "Please review the diff above and populate the structured response \
+         fields (verdict, summary, findings) as specified in the system prompt.\n",
     );
 
     msg
@@ -277,205 +326,9 @@ fn build_user_message(
 
 // ─── Unit tests ───────────────────────────────────────────────────────────────
 
+// ─── Unit tests ─────────────────────────────────────────────────────────────
+// Tests extracted to prompt_tests.rs to keep this file under the 500-line cap.
+
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn sample_meta() -> ReviewPrMeta {
-        ReviewPrMeta {
-            title: "Add authentication".to_string(),
-            author: "alice".to_string(),
-            url: "https://github.com/acme/backend/pull/42".to_string(),
-        }
-    }
-
-    fn empty_context() -> ReviewContext {
-        ReviewContext::default()
-    }
-
-    #[test]
-    fn system_prompt_contains_policy() {
-        let prompt = reviewer_system_prompt();
-        assert!(
-            prompt.contains("default verdict is APPROVE"),
-            "system prompt must state APPROVE-default policy"
-        );
-        assert!(
-            prompt.contains("REQUEST_CHANGES requires ALL THREE"),
-            "system prompt must specify the REQUEST_CHANGES gate"
-        );
-        assert!(
-            prompt.contains("BLOCK"),
-            "system prompt must describe the BLOCK tier"
-        );
-        assert!(
-            prompt.contains("\"verdict\""),
-            "system prompt must include the JSON output schema"
-        );
-    }
-
-    /// Regression test: a `bedrock/`-prefixed reviewer_model must be stripped
-    /// before being set on `LlmRequest.model`.
-    ///
-    /// Why: guards against Bug 1 regression — BedrockProvider receives the
-    /// prefixed id as the Converse model parameter, causing HTTP 400.
-    /// What: passes `bedrock/<id>` to `build_review_prompt` and asserts
-    /// `LlmRequest.model` is the bare `<id>`.
-    /// Test: this test itself; no network calls.
-    #[test]
-    fn build_review_prompt_strips_bedrock_prefix() {
-        let req = build_review_prompt(
-            "acme",
-            "backend",
-            &sample_meta(),
-            "+fn x() {}",
-            &empty_context(),
-            "bedrock/us.anthropic.claude-sonnet-4-6",
-        );
-        assert_eq!(
-            req.model, "us.anthropic.claude-sonnet-4-6",
-            "bedrock/ prefix must be stripped from LlmRequest.model"
-        );
-    }
-
-    /// Regression test: an `openrouter/`-prefixed model must also be stripped.
-    ///
-    /// Why: same Bug 1 pattern; OpenRouter API does not accept the routing prefix.
-    /// What: passes `openrouter/<id>` and asserts the bare id is used.
-    /// Test: this test itself; no network calls.
-    #[test]
-    fn build_review_prompt_strips_openrouter_prefix() {
-        let req = build_review_prompt(
-            "acme",
-            "backend",
-            &sample_meta(),
-            "+fn x() {}",
-            &empty_context(),
-            "openrouter/openai/gpt-5.4-mini-20260317",
-        );
-        assert_eq!(
-            req.model, "openai/gpt-5.4-mini-20260317",
-            "openrouter/ prefix must be stripped from LlmRequest.model"
-        );
-    }
-
-    #[test]
-    fn build_review_prompt_includes_diff() {
-        let diff = "+fn hello() { println!(\"hi\"); }\n";
-        let req = build_review_prompt(
-            "acme",
-            "backend",
-            &sample_meta(),
-            diff,
-            &empty_context(),
-            "openai/gpt-5.4-mini-20260317",
-        );
-        assert_eq!(req.model, "openai/gpt-5.4-mini-20260317");
-        assert_eq!(req.messages.len(), 1);
-        let content = &req.messages[0].content;
-        assert!(
-            content.contains("fn hello"),
-            "user message must include the diff"
-        );
-        assert!(
-            content.contains("acme/backend"),
-            "user message must include owner/repo"
-        );
-        assert!(
-            content.contains("Add authentication"),
-            "user message must include PR title"
-        );
-        assert!((req.temperature - REVIEWER_TEMPERATURE).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn prompt_includes_context_blocks() {
-        use crate::integrations::search_client::SearchResult;
-
-        let context = ReviewContext {
-            search_results: vec![SearchResult {
-                file: "src/auth.rs".to_string(),
-                snippet: Some("pub fn verify() {}".to_string()),
-                score: 0.9,
-                start_line: Some(10),
-                end_line: Some(12),
-            }],
-            complexity_hotspots: vec![ComplexityHotspot {
-                file: "src/auth.rs".to_string(),
-                function_name: Some("verify".to_string()),
-                cyclomatic: 12,
-                cognitive: 8,
-            }],
-            smells: vec![Smell {
-                file: "src/auth.rs".to_string(),
-                category: "long_method".to_string(),
-                severity: "medium".to_string(),
-                line: Some(20),
-            }],
-        };
-
-        let req = build_review_prompt(
-            "acme",
-            "repo",
-            &sample_meta(),
-            "+fn foo() {}",
-            &context,
-            "openai/gpt-5.4-mini-20260317",
-        );
-        let content = &req.messages[0].content;
-        assert!(
-            content.contains("Related code"),
-            "user message must include search context section"
-        );
-        assert!(
-            content.contains("pub fn verify"),
-            "user message must include search snippet"
-        );
-        assert!(
-            content.contains("Complexity hotspots"),
-            "user message must include hotspot section"
-        );
-        assert!(
-            content.contains("Code smells"),
-            "user message must include smells section"
-        );
-    }
-
-    #[test]
-    fn prompt_empty_context_omits_sections() {
-        let req = build_review_prompt(
-            "o",
-            "r",
-            &sample_meta(),
-            "+fn x() {}",
-            &empty_context(),
-            "openai/gpt-5.4-nano-20260317",
-        );
-        let content = &req.messages[0].content;
-        assert!(
-            !content.contains("Related code"),
-            "empty context must not include search section"
-        );
-        assert!(
-            !content.contains("Complexity hotspots"),
-            "empty context must not include hotspot section"
-        );
-    }
-
-    #[test]
-    fn prompt_local_diff_mode_no_pr_metadata() {
-        // In --local-diff mode, pr_meta has empty fields.
-        let meta = ReviewPrMeta::default();
-        let req = build_review_prompt(
-            "local",
-            "local",
-            &meta,
-            "+fn local_fn() {}",
-            &empty_context(),
-            "openai/gpt-5.4-mini-20260317",
-        );
-        let content = &req.messages[0].content;
-        // Must still include the diff.
-        assert!(content.contains("local_fn"));
-    }
-}
+#[path = "prompt_tests.rs"]
+mod tests;
